% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\name{mta}
\alias{mta}
\title{Mean Target Achievement (MTA) Function}
\usage{
mta(data = list(), target, plot = TRUE)
}
\arguments{
\item{data}{data is a dataframe with three columns: feature, ai, and pi. feature is the name of the conservation features (e.g. ecoregions, habitats, species), ai is the total area/amount of conservation features, and pi is the protected area/amount of conservation features.}

\item{target}{target is the conservation target as a value between 0 and 1 (0 is zero protection and 1 is 100\% protection).}

\item{plot}{plots the conservation target and the protected amount of conservation features ordered from low to high. Defaults to TRUE.}
}
\value{
\tabular{rlll}{\tab \code{mta}   \tab Mean Target Achievement value (0 is no achievement and 1 is 100\% achievement of conservation target) \cr \tab \code{target} \tab conservation target \cr \tab \code{N} \tab number of conservation features \cr \tab \code{proportion_protected}  \tab protected proportion of conservation features as a value from 0 to 1, sorted from low to high}
}
\description{
The mean target achievement (MTA) function determines the degree of conservation target achievement as a value between 0 and 1.
}
\examples{
# Generate input data
feature <- paste("Ecoregion",1:10)  #conservation feature names
ai      <- c(41,223,1053,520,230,303,343,2684,6507,1010)  #total amount of conservation features
pi      <- c(0,53,282,237,70,5,123,606,2695,496)  #protected amount of conservation features
data    <- data.frame(feature,ai,pi)

# Run the mta function for conservation target 0.5 (50\% protection of each feature)
mta(data,0.5,plot=FALSE)
}
\keyword{conservation}
\keyword{protection}
\keyword{target}
